likelihood ratio test

+++ src/main/java/plugins/fr/univ_nantes/ec_clem/error/likelihood_ratio/LikelihoodRatioTest.java
@@ -9,13 +9,9 @@ public class LikelihoodRatioTest {
     @Inject
     public LikelihoodRatioTest() {}
 
-    public double test(int degreesOfFreedom, double nestedModelLogLikelihood, double generalModelLogLikelihood) {
+    public double test(double degreesOfFreedom, double nestedModelLogLikelihood, double generalModelLogLikelihood) {
         ChiSquaredDistribution chi2 = new ChiSquaredDistribution(degreesOfFreedom);
         double likelihoodRatio = 2d * (generalModelLogLikelihood - nestedModelLogLikelihood);
-//        System.out.println(nestedModelLogLikelihood);
-//        System.out.println(generalModelLogLikelihood);
-//        System.out.println(likelihoodRatio);
-//        System.out.println(chi2.cumulativeProbability(likelihoodRatio));
         return 1d - chi2.cumulativeProbability(likelihoodRatio);
     }
 }
